+++ src/angular-vs-repeat.js
@@ -1,78 +1,84 @@
 /**
  * Copyright Kamil Pękala http://github.com/kamilkp
- * Angular Virtual Scroll Repeat v2.0.0 2018/03/15
+ * Angular Virtual Scroll Repeat v2.0.0 2018/03/16
  */
 
 /* global console, setTimeout, module */
 
 ((window, angular) => {
-  // DESCRIPTION:
-  // vsRepeat directive stands for Virtual Scroll Repeat. It turns a standard ngRepeated set of elements in a scrollable container
-  // into a component, where the user thinks he has all the elements rendered and all he needs to do is scroll (without any kind of
-  // pagination - which most users loath) and at the same time the browser isn't overloaded by that many elements/angular bindings etc.
-  // The directive renders only so many elements that can fit into current container's clientHeight/clientWidth.
+  /**
+   * DESCRIPTION:
+   * vsRepeat directive stands for Virtual Scroll Repeat. It turns a standard ngRepeated set of elements in a scrollable container
+   * into a component, where the user thinks he has all the elements rendered and all he needs to do is scroll (without any kind of
+   * pagination - which most users loath) and at the same time the browser isn't overloaded by that many elements/angular bindings etc.
+   * The directive renders only so many elements that can fit into current container's clientHeight/clientWidth.
 
-  // LIMITATIONS:
-  // - current version only supports an Array as a right-hand-side object for ngRepeat
-  // - all rendered elements must have the same height/width or the sizes of the elements must be known up front
+   * LIMITATIONS:
+   * - current version only supports an Array as a right-hand-side object for ngRepeat
+   * - all rendered elements must have the same height/width or the sizes of the elements must be known up front
 
-  // USAGE:
-  // In order to use the vsRepeat directive you need to place a vs-repeat attribute on a direct parent of an element with ng-repeat
-  // example:
-  // <div vs-repeat>
-  //      <div ng-repeat="item in someArray">
-  //          <!-- content -->
-  //      </div>
-  // </div>
-  //
-  // or:
-  // <div vs-repeat>
-  //      <div ng-repeat-start="item in someArray">
-  //          <!-- content -->
-  //      </div>
-  //      <div>
-  //         <!-- something in the middle -->
-  //      </div>
-  //      <div ng-repeat-end>
-  //          <!-- content -->
-  //      </div>
-  // </div>
-  //
-  // You can also measure the single element's height/width (including all paddings and margins), and then speficy it as a value
-  // of the attribute 'vs-repeat'. This can be used if one wants to override the automatically computed element size.
-  // example:
-  // <div vs-repeat="50"> <!-- the specified element height is 50px -->
-  //      <div ng-repeat="item in someArray">
-  //          <!-- content -->
-  //      </div>
-  // </div>
-  //
-  // IMPORTANT!
-  //
-  // - the vsRepeat directive must be applied to a direct parent of an element with ngRepeat
-  // - the value of vsRepeat attribute is the single element's height/width measured in pixels. If none provided, the directive
-  //      will compute it automatically
+   * USAGE:
+   * In order to use the vsRepeat directive you need to place a vs-repeat attribute on a direct parent of an element with ng-repeat
+   * example:
+   * <div vs-repeat="options">
+   *      <div ng-repeat="item in someArray">
+   *          <!-- content -->
+   *      </div>
+   * </div>
 
-  // OPTIONAL PARAMETERS (attributes):
-  // vs-repeat-container="selector" - selector for element containing ng-repeat. (defaults to the current element)
-  // vs-scroll-parent="selector" - selector to the scrollable container. The directive will look for a closest parent matching
-  //                              the given selector (defaults to the current element)
-  // vs-horizontal - stack repeated elements horizontally instead of vertically
-  // vs-offset-before="value" - top/left offset in pixels (defaults to 0)
-  // vs-offset-after="value" - bottom/right offset in pixels (defaults to 0)
-  // vs-excess="value" - an integer number representing the number of elements to be rendered outside of the current container's viewport
-  //                      (defaults to 2)
-  // vs-size - a property name of the items in collection that is a number denoting the element size (in pixels)
-  // vs-autoresize - use this attribute without vs-size and without specifying element's size. The automatically computed element style will
-  //              readjust upon window resize if the size is dependable on the viewport size
-  // vs-scrolled-to-end="callback" - callback will be called when the last item of the list is rendered
-  // vs-scrolled-to-end-offset="integer" - set this number to trigger the scrolledToEnd callback n items before the last gets rendered
-  // vs-scrolled-to-beginning="callback" - callback will be called when the first item of the list is rendered
-  // vs-scrolled-to-beginning-offset="integer" - set this number to trigger the scrolledToBeginning callback n items before the first gets rendered
+   * or:
+   * <div vs-repeat="options">
+   *      <div ng-repeat-start="item in someArray">
+   *          <!-- content -->
+   *      </div>
+   *      <div>
+   *         <!-- something in the middle -->
+   *      </div>
+   *      <div ng-repeat-end>
+   *          <!-- content -->
+   *      </div>
+   * </div>
 
-  // EVENTS:
-  // - 'vsRepeatTrigger' - an event the directive listens for to manually trigger reinitialization
-  // - 'vsRepeatReinitialized' - an event the directive emits upon reinitialization done
+   * You can also measure the single element's height/width (including all paddings and margins), and then speficy it as a value
+   * of the option's `size` property. This can be used if one wants to override the automatically computed element size.
+   * example:
+   * <div vs-repeat="{size: 50}"> <!-- the specified element height is 50px -->
+   *      <div ng-repeat="item in someArray">
+   *          <!-- content -->
+   *      </div>
+   * </div>
+
+   * IMPORTANT!
+
+   * - the vsRepeat directive must be applied to a direct parent of an element with ngRepeat
+   * - the value of vsRepeat attribute is the single element's height/width measured in pixels. If none provided, the directive
+   *      will compute it automatically
+
+   * OPTIONAL PARAMETERS (attributes):
+   * vs-repeat-container="selector" - selector for element containing ng-repeat. (defaults to the current element)
+
+   * OPTIONS:
+   * Options shall be passed as an object to the `vs-repeat` attribute e.g.: `<div vs-repeat="{scrollParent: 'window', size: 20}"></div>`
+   *
+   * Available options:
+   * `horizontal` - stack repeated elements horizontally instead of vertically
+   * `offset-before` - top/left offset in pixels (defaults to 0)
+   * `offset-after` - bottom/right offset in pixels (defaults to 0)
+   * `scroll-margin` - how many pixels ahead should elements be rendered while scrolling
+   * `latch` - if true, elements will be rendered gradually but won't be removed when scrolled away (defaults to false)
+
+   * `size` - a property name of the items in collection that is a number denoting the element size (in pixels)
+   * `autoresize` - use this attribute without vs-size and without specifying element's size. The automatically computed element style will
+   *              readjust upon window resize if the size is dependable on the viewport size
+   * `scrolled-to-end` - callback will be called when the last item of the list is rendered
+   * `scrolled-to-end-offset` - set this number to trigger the scrolledToEnd callback n items before the last gets rendered
+   * `scrolled-to-beginning` - callback will be called when the first item of the list is rendered
+   * `scrolled-to-beginning-offset` - set this number to trigger the scrolledToBeginning callback n items before the first gets rendered
+
+   * EVENTS:
+   * - `vsRepeatTrigger` - an event the directive listens for to manually trigger reinitialization
+   * - `vsRepeatReinitialized` - an event the directive emits upon reinitialization done
+   */
 
   let closestElement = angular.element.prototype.closest;
 
@@ -152,12 +158,12 @@
     throw new Error('angular-vs-repeat: no ng-repeat directive on a child element');
   }
 
-  function printDeprecationWarning(message) {
-    console.warn(`vs-repeat deprecation: ${message}`);
+  function printDeprecationWarning($element, message) {
+    console.warn(`vs-repeat deprecation: ${message}`, $element[0]);
   }
 
-  function attrDeprecated(attrname) {
-    printDeprecationWarning(`${attrname} attribute is deprecated. Use vs-repeat-options <link> instead.`);
+  function attrDeprecated(attrname, $element) {
+    printDeprecationWarning($element, `${attrname} attribute is deprecated. Pass the options object to vs-repeat attribute instead https://github.com/kamilkp/angular-vs-repeat#options`);
   }
 
   const defaultOptions = {
@@ -202,13 +208,9 @@
           'vsScrollMargin',
         ].forEach((attrname) => {
           if (attrname in compileAttrs) {
-            attrDeprecated(attrname);
+            attrDeprecated(attrname, compileElement);
           }
         });
-
-        if (compileAttrs.vsRepeat) {
-          attrDeprecated('vsRepeat');
-        }
 
         const [originalNgRepeatAttr, ngRepeatExpression, isNgRepeatStart] = analyzeNgRepeatUsage(ngRepeatChild);
 
@@ -240,7 +242,7 @@
             $scope.vsRepeat = {
               options: {
                 ...defaultOptions,
-                ...('vsRepeatOptions' in $attrs ? $scope.$eval($attrs.vsRepeatOptions) : {}),
+                ...($scope.$eval($attrs.vsRepeat) ?? {}),
               },
             };
 
@@ -412,12 +414,6 @@
             function scrollHandler() {
               if (updateInnerCollection()) {
                 $scope.$digest();
-
-                const expectedSize = $scope.vsRepeat.sizesCumulative[originalLength];
-
-                if (expectedSize !== repeatContainer[0][scrollSize]) {
-                  console.warn('vsRepeat: size mismatch. Expected size ' + expectedSize + 'px whereas actual size is ' + repeatContainer[0][scrollSize] + 'px. Fix vsSize on element:', $element[0]);
-                }
               }
             }
 
@@ -532,7 +528,11 @@
                 return binaryFind(array, threshold, m, b, d + 1);
               }
 
-              return [a, b, d];
+              return [
+                threshold > array[b] ? b : a,
+                threshold < array[a] ? a : b,
+                d,
+              ];
             }
 
             function updateInnerCollection() {
@@ -556,6 +556,8 @@
                 __startIndex = 0;
                 __endIndex = 1;
               } else {
+                _warnMismatch();
+
                 const relativeScroll = $scrollPosition - options.offsetBefore - scrollOffset;
 
                 ([__startIndex] = binaryFind($scope.vsRepeat.sizesCumulative, relativeScroll - options.scrollMargin));
@@ -628,6 +630,23 @@
               }
 
               return digestRequired;
+            }
+
+            function _warnMismatch() {
+              $scope.$$postDigest(() => {
+                window.requestAnimationFrame(() => {
+                  const expectedSize = $scope.vsRepeat.sizesCumulative[originalLength];
+                  const compStyle = window.getComputedStyle(repeatContainer[0]);
+                  const paddings = options.horizontal ?
+                    ['paddingLeft', 'paddingRight']
+                    : ['paddingTop', 'paddingBottom'];
+                  const containerSize = repeatContainer[0][scrollSize] - paddings.reduce((acc, prop) => acc + Number(compStyle[prop].slice(0, -2)), 0);
+
+                  if (repeatContainer[0][scrollSize] && expectedSize !== containerSize) {
+                    console.warn('vsRepeat: size mismatch. Expected size ' + expectedSize + 'px whereas actual size is ' + containerSize + 'px. Fix vsSize on element:', $element[0]);
+                  }
+                });
+              });
             }
           },
         };